Refactor bff and publicapi adapters packages
