+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/infrastructure/api/adapters/bff/inputs/CreateOrUpdateMissionDataInput.kt
@@ -1,4 +1,4 @@
-package fr.gouv.cacem.monitorenv.infrastructure.api.adapters.inputs
+package fr.gouv.cacem.monitorenv.infrastructure.api.adapters.bff.inputs
 
 import fr.gouv.cacem.monitorenv.domain.entities.controlResources.ControlUnitEntity
 import fr.gouv.cacem.monitorenv.domain.entities.missions.EnvActionEntity
@@ -8,7 +8,7 @@ import fr.gouv.cacem.monitorenv.domain.entities.missions.MissionTypeEnum
 import org.locationtech.jts.geom.MultiPolygon
 import java.time.ZonedDateTime
 
-data class CreateOrUpdatePublicMissionDataInput(
+data class CreateOrUpdateMissionDataInput(
     val id: Int? = null,
     val missionTypes: List<MissionTypeEnum>,
     val controlUnits: List<ControlUnitEntity> = listOf(),
@@ -23,11 +23,13 @@ data class CreateOrUpdatePublicMissionDataInput(
     val missionSource: MissionSourceEnum,
     val isClosed: Boolean,
     val envActions: List<EnvActionEntity>? = null,
-    val hasMissionOrder: Boolean,
-    val isUnderJdp: Boolean,
-    val isGeometryComputedFromControls: Boolean
+    val hasMissionOrder: Boolean? = false,
+    val isUnderJdp: Boolean? = false,
 ) {
     fun toMissionEntity(): MissionEntity {
+        val hasMissionOrder = this.hasMissionOrder ?: false
+        val isUnderJdp = this.isUnderJdp ?: false
+
         return MissionEntity(
             id = this.id,
             missionTypes = this.missionTypes,
@@ -44,9 +46,9 @@ data class CreateOrUpdatePublicMissionDataInput(
             isDeleted = false,
             missionSource = this.missionSource,
             envActions = this.envActions,
-            hasMissionOrder = this.hasMissionOrder,
-            isUnderJdp = this.isUnderJdp,
-            isGeometryComputedFromControls = this.isGeometryComputedFromControls
+            hasMissionOrder = hasMissionOrder,
+            isUnderJdp = isUnderJdp,
+            isGeometryComputedFromControls = false
         )
     }
 }
