+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/infrastructure/api/adapters/bff/outputs/MissingParameterApiError.kt
@@ -1,0 +1,3 @@
+package fr.gouv.cacem.monitorenv.infrastructure.api.adapters.bff.outputs
+
+data class MissingParameterApiError(val error: String)
