--- backend/src/main/kotlin/fr/gouv/cacem/monitorenv/infrastructure/api/adapters/outputs/MissingParameterApiError.kt
@@ -1,3 +1,0 @@
-package fr.gouv.cacem.monitorenv.infrastructure.api.adapters.outputs
-
-data class MissingParameterApiError(val error: String)
